==> src/1. EXTRACCION.ipynb <==
import pandas as pd
from datetime import datetime
from time import sleep

from funciones.extraccion_balance import *
from funciones.extraccion_demanda import *
from funciones.extraccion_generacion import *
from funciones.extraccion_intercambios import *

### EXTRACCION DE DATOS

Para la extracción de los datos nos hemos inspirado en la propia herramienta de visualización que ofrece Red Eléctrica Española: [Aldia](https://www.ree.es/es/datos/aldia)

Datos a extraer:

- Demanda

- Balance

- Generación

- Intercambios

---

#### DEFINICIÓN DE PARÁMETROS PARA QUERY GET

**ACLARACIÓN: Para facilitar el mantenimiento y la legibilidad, hemos definido los parámetros en un archivo independiente (parametros.py)**

La API ofrece mucha información, alguna que no sabemos interpretar. 

Para intentar simplificar la cantidad de datos y su posterior legibilidad y aplicabilidad, nos hemos centrado campos concretos: 

- IDIOMA: 

    * Español

- CATEGORIAS y WIDGETS:

    Hemos contemplado las categorías requeridas por el proyecto y los widgets que tienen información territorial completa:

    * Balance
        - balance-electrico
        
    * Demanda
        - evolucion
        - perdidas-transporte
        - ire-general

    * Generación
        - estructura-generacion

    * Intercambios
        - francia-frontera
        - portugal-frontera
        - marruecos-frontera
        - andorra-frontera
        - enlace-baleares
    
    INFORMACIÓN TERRITORIAL
    
    * Nacional 
    * Peninsular
    * Ceuta
    * Melilla
    * Canarias
    * Baleares
    * CCAA

---

### SCRIPTS

**ACLARACIÓN: Para facilitar el mantenimiento y la legibilidad, hemos definido los scripts de extracción en un archivo independiente (extraccion_xxx.py)**

La solicitud de cada extracción se ha definido con una función que recibe fecha de inicio y fecha de fin. Esto hace que el mismo código se pueda reutilizar en diferentes intervalos de tiempo. 

Se ha definido en intervalos de 1 año cada vez con la idea de no saturar el servidor de REE ni de provocar tiempos de espera demasiado altos.

Extraemos datos desde el 2010 hasta el día de hoy

---

hoy = datetime.now() 
inicio = "2010-01-01T00:00"
fin = hoy.strftime("%Y-%m-%dT23:59")

fechas = { inicio : "2010-12-31T23:59", 
          "2011-01-01T00:00" : "2011-12-31T23:59", 
          "2012-01-01T00:00" : "2012-12-31T23:59", 
          "2013-01-01T00:00" : "2013-12-31T23:59", 
          "2014-01-01T00:00" : "2014-12-31T23:59", 
          "2015-01-01T00:00" : "2015-12-31T23:59", 
          "2016-01-01T00:00" : "2016-12-31T23:59", 
          "2017-01-01T00:00" : "2017-12-31T23:59", 
          "2018-01-01T00:00" : "2018-12-31T23:59", 
          "2019-01-01T00:00" : "2019-12-31T23:59", 
          "2020-01-01T00:00" : "2020-12-31T23:59", 
          "2021-01-01T00:00" : "2021-12-31T23:59", 
          "2022-01-01T00:00" : "2022-12-31T23:59", 
          "2023-01-01T00:00" : "2023-12-31T23:59", 
          "2024-01-01T00:00" : fin}

#### EXTRACCION BALANCE HISTORICO 2020/ACTUALIDAD

---

df_balance_historico = pd.DataFrame()

for key, value in fechas.items():
    df_balance = extrae_balance(key, value)
    df_balance_historico = pd.concat([df_balance_historico, df_balance])
    sleep(1)

df_balance_historico.to_csv("../data/raw/DF_BALANCE_HISTORICO_10_25.csv", index=False, sep=",")

#### EXTRACCION DEMANDA HISTORICO 2020/ACTUALIDAD

---

df_demanda_historico = pd.DataFrame()

for key, value in fechas.items():
    df_demanda = extrae_demanda(key, value)
    df_demanda_historico = pd.concat([df_demanda_historico, df_demanda])
    sleep(1)

df_demanda_historico.to_csv("../data/raw/DF_DEMANDA_HISTORICO_10_25.csv", index=False, sep=",")

#### EXTRACCION GENERACION HISTORICO 2020/ACTUALIDAD

---

df_generacion_historico = pd.DataFrame()

for key, value in fechas.items():
    df_generacion = extrae_generacion(key, value)
    df_generacion_historico = pd.concat([df_generacion_historico, df_generacion])
    sleep(1)

df_generacion_historico.to_csv("../data/raw/DF_GENERACION_HISTORICO_10_25.csv", index=False, sep=",")

#### EXTRACCION INTERCAMBIOS HISTORICO 2020/ACTUALIDAD

---

df_intercambios_historico = pd.DataFrame()

for key, value in fechas.items():
    df_intercambios = extrae_intercambios(key, value)
    df_intercambios_historico = pd.concat([df_intercambios_historico, df_intercambios])
    sleep(1)

df_intercambios_historico.to_csv("../data/raw/DF_INTERCAMBIOS_HISTORICO_10_25.csv", index=False, sep=",")
