==> src/3. CALCULO BALANCE Y DEMANDA.ipynb <==
import pandas as pd
from datetime import datetime

from funciones.calculo_balance_demanda import *

hoy = datetime.now() 

inicio = "2020-01-01"
fin = hoy.strftime("%Y-%m-%d")

print(hoy)
print(inicio)
print(fin)

zonas = ["nacional", "peninsular", "ceuta", "melilla", "canarias", "baleares"]

## CALCULOS BALANCE Y DEMANDA ENERGETICA

#### Segun los parámetros que pasemos a las funciones podremos obtener un dataframe con el demanda y/o la balance energéticos para zona e intervalo de tiempo determinados, agrupado por días.

---

### CÁLCULO BALANCE ENERGÉTICO:

Generación (todas las tecnologías)+ Saldo almacenamiento + Saldo intercambios internacionales

balance_energetico_total(inicio, fin, "nacional")

---

### CÁLCULO DEMANDA ENERGÉTICA:

Generación (todas las tecnologías) + pérdidas de transporte + Saldo Intercambios Internacionales (positivo o negativo) 

demanda_energetica_total(inicio, fin, "nacional")
