==> src/5. ACTUALIZACION_DATAFRAMES.ipynb <==
### ACTUALIZACION DATAFRAMES Y BBDD

- actualiza_dfs():

    Para mantener los datos actualizados, extraemos, limpiamos y sustituimos outliers cada cierto tiempo. Por ello, algunos de los datos de la columna valor_MWh(provisionales) cambian por datos definitivos. Esto duplica registros así que hemos tenido que evaluar varias veces cada dataframe actualizado (concatenado dataframe "base" con dataframe "nuevo").



- actualiza_bbdd():

    Los dataframes tiene una columna única (ID) que se crea concatenando fecha de medición + UUID. El problema de este método es que al reejecutar el código, el ID cambia, lo que nos ha traído problemas para poder actualizar los datos en la BBDD valorando si ya existían o no. Por este motivo, decidimos finalmente sobreescribrir la BBDD con los nuevos dataframes concatenados y sin duplicados. 

    Además, durante las pruebas de la mejor opción a implementar vimos que SQLAlchemy no conectaba correctamente con la BBDD, lo que nos obligó a usar mysql.connector también en la misma función.

---

import pandas as pd
from funciones.actualiza_dataframes import *
from database.actualiza_bbdd import *

df_balance_historico_actualizado, \
df_demanda_historico_actualizado, \
df_generacion_historico_actualizado, \
df_intercambios_historico_actualizado = actualiza_dfs()

# df_balance_historico_actualizado = pd.read_csv("../data/processed/DF_BALANCE_10_25_LIMPIO.csv")
# df_demanda_historico_actualizado = pd.read_csv("../data/processed/DF_DEMANDA_10_25_LIMPIO.csv")
# df_generacion_historico_actualizado = pd.read_csv("../data/processed/DF_GENERACION_10_25_LIMPIO.csv")
# df_intercambios_historico_actualizado = pd.read_csv("../data/processed/DF_INTERCAMBIOS_10_25_LIMPIO.csv")

actualiza_bbdd(df_balance_historico_actualizado, 
               df_demanda_historico_actualizado, 
               df_generacion_historico_actualizado, 
               df_intercambios_historico_actualizado)

df_balance_historico_actualizado.to_csv("../data/processed/DF_BALANCE_10_25_LIMPIO.csv", index=False, sep=",")
df_demanda_historico_actualizado.to_csv("../data/processed/DF_DEMANDA_10_25_LIMPIO.csv", index=False, sep=",")
df_generacion_historico_actualizado.to_csv("../data/processed/DF_GENERACION_10_25_LIMPIO.csv", index=False, sep=",")
df_intercambios_historico_actualizado.to_csv("../data/processed/DF_INTERCAMBIOS_10_25_LIMPIO.csv", index=False, sep=",")

# print para comprobar registros con BBDD manualmente

print(len(df_balance_historico_actualizado))
print(len(df_demanda_historico_actualizado))
print(len(df_generacion_historico_actualizado))
print(len(df_intercambios_historico_actualizado))
